+++ table/block_based/block_based_table_reader.cc
@@ -583,6 +583,7 @@ Status BlockBasedTable::Open(
   ro.io_timeout = read_options.io_timeout;
   ro.rate_limiter_priority = read_options.rate_limiter_priority;
   ro.verify_checksums = read_options.verify_checksums;
+  ro.io_activity = read_options.io_activity;
 
   // prefetch both index and filters, down to all partitions
   const bool prefetch_all = prefetch_index_and_filter_in_cache || level == 0;
@@ -2231,7 +2232,8 @@ Status BlockBasedTable::MultiGetFilter(const ReadOptions& read_options,
   return Status::OK();
 }
 
-Status BlockBasedTable::Prefetch(const Slice* const begin,
+Status BlockBasedTable::Prefetch(const ReadOptions& read_options,
+                                 const Slice* const begin,
                                  const Slice* const end) {
   auto& comparator = rep_->internal_comparator;
   UserComparatorWrapper user_comparator(comparator.user_comparator());
@@ -2241,7 +2243,7 @@ Status BlockBasedTable::Prefetch(const Slice* const begin,
   }
   BlockCacheLookupContext lookup_context{TableReaderCaller::kPrefetch};
   IndexBlockIter iiter_on_stack;
-  auto iiter = NewIndexIterator(ReadOptions(), /*need_upper_bound_check=*/false,
+  auto iiter = NewIndexIterator(read_options, /*need_upper_bound_check=*/false,
                                 &iiter_on_stack, /*get_context=*/nullptr,
                                 &lookup_context);
   std::unique_ptr<InternalIteratorBase<IndexValue>> iiter_unique_ptr;
@@ -2278,7 +2280,7 @@ Status BlockBasedTable::Prefetch(const Slice* const begin,
     DataBlockIter biter;
     Status tmp_status;
     NewDataBlockIterator<DataBlockIter>(
-        ReadOptions(), block_handle, &biter, /*type=*/BlockType::kData,
+        read_options, block_handle, &biter, /*type=*/BlockType::kData,
         /*get_context=*/nullptr, &lookup_context,
         /*prefetch_buffer=*/nullptr, /*for_compaction=*/false,
         /*async_read=*/false, tmp_status);
@@ -2298,11 +2300,10 @@ Status BlockBasedTable::VerifyChecksum(const ReadOptions& read_options,
   // Check Meta blocks
   std::unique_ptr<Block> metaindex;
   std::unique_ptr<InternalIterator> metaindex_iter;
-  ReadOptions ro;
-  s = ReadMetaIndexBlock(ro, nullptr /* prefetch buffer */, &metaindex,
-                         &metaindex_iter);
+  s = ReadMetaIndexBlock(read_options, nullptr /* prefetch buffer */,
+                         &metaindex, &metaindex_iter);
   if (s.ok()) {
-    s = VerifyChecksumInMetaBlocks(metaindex_iter.get());
+    s = VerifyChecksumInMetaBlocks(read_options, metaindex_iter.get());
     if (!s.ok()) {
       return s;
     }
@@ -2409,7 +2410,7 @@ BlockType BlockBasedTable::GetBlockTypeForMetaBlockByName(
 }
 
 Status BlockBasedTable::VerifyChecksumInMetaBlocks(
-    InternalIteratorBase<Slice>* index_iter) {
+    const ReadOptions& read_options, InternalIteratorBase<Slice>* index_iter) {
   Status s;
   for (index_iter->SeekToFirst(); index_iter->Valid(); index_iter->Next()) {
     s = index_iter->status();
@@ -2425,14 +2426,14 @@ Status BlockBasedTable::VerifyChecksumInMetaBlocks(
       // Unfortunate special handling for properties block checksum w/
       // global seqno
       std::unique_ptr<TableProperties> table_properties;
-      s = ReadTablePropertiesHelper(ReadOptions(), handle, rep_->file.get(),
+      s = ReadTablePropertiesHelper(read_options, handle, rep_->file.get(),
                                     nullptr /* prefetch_buffer */, rep_->footer,
                                     rep_->ioptions, &table_properties,
                                     nullptr /* memory_allocator */);
     } else {
       s = BlockFetcher(
               rep_->file.get(), nullptr /* prefetch buffer */, rep_->footer,
-              ReadOptions(), handle, &contents, rep_->ioptions,
+              read_options, handle, &contents, rep_->ioptions,
               false /* decompress */, false /*maybe_compressed*/,
               GetBlockTypeForMetaBlockByName(meta_block_name),
               UncompressionDict::GetEmptyDict(), rep_->persistent_cache_options)
@@ -2544,7 +2545,8 @@ uint64_t BlockBasedTable::GetApproximateDataSize() {
   return rep_->footer.metaindex_handle().offset();
 }
 
-uint64_t BlockBasedTable::ApproximateOffsetOf(const Slice& key,
+uint64_t BlockBasedTable::ApproximateOffsetOf(const ReadOptions& read_options,
+                                              const Slice& key,
                                               TableReaderCaller caller) {
   uint64_t data_size = GetApproximateDataSize();
   if (UNLIKELY(data_size == 0)) {
@@ -2558,6 +2560,7 @@ uint64_t BlockBasedTable::ApproximateOffsetOf(const Slice& key,
   IndexBlockIter iiter_on_stack;
   ReadOptions ro;
   ro.total_order_seek = true;
+  ro.io_activity = read_options.io_activity;
   auto index_iter =
       NewIndexIterator(ro, /*disable_prefix_seek=*/true,
                        /*input_iter=*/&iiter_on_stack, /*get_context=*/nullptr,
@@ -2586,7 +2589,8 @@ uint64_t BlockBasedTable::ApproximateOffsetOf(const Slice& key,
                                static_cast<double>(rep_->file_size));
 }
 
-uint64_t BlockBasedTable::ApproximateSize(const Slice& start, const Slice& end,
+uint64_t BlockBasedTable::ApproximateSize(const ReadOptions& read_options,
+                                          const Slice& start, const Slice& end,
                                           TableReaderCaller caller) {
   assert(rep_->internal_comparator.Compare(start, end) <= 0);
 
@@ -2603,6 +2607,7 @@ uint64_t BlockBasedTable::ApproximateSize(const Slice& start, const Slice& end,
   IndexBlockIter iiter_on_stack;
   ReadOptions ro;
   ro.total_order_seek = true;
+  ro.io_activity = read_options.io_activity;
   auto index_iter =
       NewIndexIterator(ro, /*disable_prefix_seek=*/true,
                        /*input_iter=*/&iiter_on_stack, /*get_context=*/nullptr,
@@ -2654,9 +2659,9 @@ bool BlockBasedTable::TEST_IndexBlockInCache() const {
 }
 
 Status BlockBasedTable::GetKVPairsFromDataBlocks(
-    std::vector<KVPairBlock>* kv_pair_blocks) {
+    const ReadOptions& read_options, std::vector<KVPairBlock>* kv_pair_blocks) {
   std::unique_ptr<InternalIteratorBase<IndexValue>> blockhandles_iter(
-      NewIndexIterator(ReadOptions(), /*need_upper_bound_check=*/false,
+      NewIndexIterator(read_options, /*need_upper_bound_check=*/false,
                        /*input_iter=*/nullptr, /*get_context=*/nullptr,
                        /*lookup_contex=*/nullptr));
 
@@ -2677,7 +2682,7 @@ Status BlockBasedTable::GetKVPairsFromDataBlocks(
     std::unique_ptr<InternalIterator> datablock_iter;
     Status tmp_status;
     datablock_iter.reset(NewDataBlockIterator<DataBlockIter>(
-        ReadOptions(), blockhandles_iter->value().handle,
+        read_options, blockhandles_iter->value().handle,
         /*input_iter=*/nullptr, /*type=*/BlockType::kData,
         /*get_context=*/nullptr, /*lookup_context=*/nullptr,
         /*prefetch_buffer=*/nullptr, /*for_compaction=*/false,
@@ -2723,7 +2728,8 @@ Status BlockBasedTable::DumpTable(WritableFile* out_file) {
                 "--------------------------------------\n";
   std::unique_ptr<Block> metaindex;
   std::unique_ptr<InternalIterator> metaindex_iter;
-  ReadOptions ro;
+  // TODO: plumb Env::IOActivity
+  const ReadOptions ro;
   Status s = ReadMetaIndexBlock(ro, nullptr /* prefetch_buffer */, &metaindex,
                                 &metaindex_iter);
   if (s.ok()) {
@@ -2779,7 +2785,7 @@ Status BlockBasedTable::DumpTable(WritableFile* out_file) {
   if (rep_->uncompression_dict_reader) {
     CachableEntry<UncompressionDict> uncompression_dict;
     s = rep_->uncompression_dict_reader->GetOrReadUncompressionDictionary(
-        nullptr /* prefetch_buffer */, false /* no_io */,
+        nullptr /* prefetch_buffer */, ro, false /* no_io */,
         false, /* verify_checksums */
         nullptr /* get_context */, nullptr /* lookup_context */,
         &uncompression_dict);
@@ -2797,7 +2803,7 @@ Status BlockBasedTable::DumpTable(WritableFile* out_file) {
   }
 
   // Output range deletions block
-  auto* range_del_iter = NewRangeTombstoneIterator(ReadOptions());
+  auto* range_del_iter = NewRangeTombstoneIterator(ro);
   if (range_del_iter != nullptr) {
     range_del_iter->SeekToFirst();
     if (range_del_iter->Valid()) {
@@ -2827,8 +2833,10 @@ Status BlockBasedTable::DumpTable(WritableFile* out_file) {
 Status BlockBasedTable::DumpIndexBlock(std::ostream& out_stream) {
   out_stream << "Index Details:\n"
                 "--------------------------------------\n";
+  // TODO: plumb Env::IOActivity
+  const ReadOptions read_options;
   std::unique_ptr<InternalIteratorBase<IndexValue>> blockhandles_iter(
-      NewIndexIterator(ReadOptions(), /*need_upper_bound_check=*/false,
+      NewIndexIterator(read_options, /*need_upper_bound_check=*/false,
                        /*input_iter=*/nullptr, /*get_context=*/nullptr,
                        /*lookup_contex=*/nullptr));
   Status s = blockhandles_iter->status();
@@ -2876,8 +2884,10 @@ Status BlockBasedTable::DumpIndexBlock(std::ostream& out_stream) {
 }
 
 Status BlockBasedTable::DumpDataBlocks(std::ostream& out_stream) {
+  // TODO: plumb Env::IOActivity
+  const ReadOptions read_options;
   std::unique_ptr<InternalIteratorBase<IndexValue>> blockhandles_iter(
-      NewIndexIterator(ReadOptions(), /*need_upper_bound_check=*/false,
+      NewIndexIterator(read_options, /*need_upper_bound_check=*/false,
                        /*input_iter=*/nullptr, /*get_context=*/nullptr,
                        /*lookup_contex=*/nullptr));
   Status s = blockhandles_iter->status();
@@ -2911,7 +2921,7 @@ Status BlockBasedTable::DumpDataBlocks(std::ostream& out_stream) {
     std::unique_ptr<InternalIterator> datablock_iter;
     Status tmp_status;
     datablock_iter.reset(NewDataBlockIterator<DataBlockIter>(
-        ReadOptions(), blockhandles_iter->value().handle,
+        read_options, blockhandles_iter->value().handle,
         /*input_iter=*/nullptr, /*type=*/BlockType::kData,
         /*get_context=*/nullptr, /*lookup_context=*/nullptr,
         /*prefetch_buffer=*/nullptr, /*for_compaction=*/false,